Remove client code from nodedef.h/cpp
