+++ src/client/client.cpp
@@ -30,6 +30,7 @@
 #include "mapnode.h"
 #include "mapsector.h"
 #include "minimap.h"
+#include "node_visuals.h"
 #include "profiler.h"
 #include "shader.h"
 #include "translation.h"
@@ -57,6 +58,7 @@
 #include "database/database-files.h"
 #include "database/database-sqlite3.h"
 
+#include <IAnimatedMesh.h>
 #include <IFileSystem.h>
 #include <json/json.h>
 
@@ -1880,7 +1882,7 @@ void Client::afterContentReceived()
 	TextureUpdateArgs tu_args;
 	tu_args.last_time_ms = porting::getTimeMs();
 	tu_args.text_base = wstrgettext("Initializing nodes");
-	m_nodedef->updateTextures(this, &tu_args);
+	fillNodeVisuals(m_nodedef, this, &tu_args);
 
 	// Start mesh update thread after setting up content definitions
 	infostream<<"- Starting mesh update thread"<<std::endl;
