+++ src/client/mesh.h
@@ -4,14 +4,15 @@
 
 #pragma once
 
+#include "irrlichttypes_bloated.h"
 #include "SColor.h"
 #include "SMaterialLayer.h"
-#include "nodedef.h"
 
 namespace scene {
 	class IAnimatedMesh;
 	class IMesh;
 	class IMeshBuffer;
+	struct SMesh;
 }
