+++ src/script/common/c_content.cpp
@@ -24,6 +24,10 @@
 #include <json/json.h>
 #include "mapgen/treegen.h"
 
+#if CHECK_CLIENT_BUILD()
+#include "client/node_visuals.h"
+#endif
+
 struct EnumString es_TileAnimationType[] =
 {
 	{TAT_NONE, "none"},
@@ -1079,8 +1083,10 @@ void push_content_features(lua_State *L, const ContentFeatures &c)
 		lua_setfield(L, -2, "mesh");
 	}
 #if CHECK_CLIENT_BUILD()
-	push_ARGB8(L, c.minimap_color);       // I know this is not set-able w/ register_node,
-	lua_setfield(L, -2, "minimap_color"); // but the people need to know!
+	if (c.visuals) {
+		push_ARGB8(L, c.visuals->minimap_color); // I know this is not set-able w/ register_node,
+		lua_setfield(L, -2, "minimap_color");    // but the people need to know!
+	}
 #endif
 	lua_pushnumber(L, c.visual_scale);
 	lua_setfield(L, -2, "visual_scale");
@@ -1093,8 +1099,12 @@ void push_content_features(lua_State *L, const ContentFeatures &c)
 		lua_pushstring(L, c.palette_name.c_str());
 		lua_setfield(L, -2, "palette_name");
 
-		push_palette(L, c.palette);
-		lua_setfield(L, -2, "palette");
+#if CHECK_CLIENT_BUILD()
+		if (c.visuals) {
+			push_palette(L, c.visuals->palette);
+			lua_setfield(L, -2, "palette");
+		}
+#endif
 	}
 	lua_pushnumber(L, c.waving);
 	lua_setfield(L, -2, "waving");
