+++ src/unittest/test_content_mapblock.cpp
@@ -13,6 +13,7 @@
 #include "client/content_mapblock.h"
 #include "client/mapblock_mesh.h"
 #include "client/meshgen/collector.h"
+#include "client/node_visuals.h"
 #include "mesh_compare.h"
 
 namespace {
@@ -27,13 +28,24 @@ public:
 		return const_cast<NodeDefManager *>(m_nodedef);
 	}
 
-	content_t registerNode(ItemDefinition itemdef, ContentFeatures nodedef) {
+	// ContentFeatures that doesn't destroy the visuals
+	struct CContentFeatures : public ContentFeatures {
+		~CContentFeatures() { visuals = nullptr; }
+	};
+
+	content_t registerNode(ItemDefinition itemdef, const CContentFeatures &nodedef) {
 		item_mgr()->registerItem(itemdef);
 		return node_mgr()->set(nodedef.name, nodedef);
 	}
 
 	void finalize() {
 		node_mgr()->resolveCrossrefs();
+
+		// Need to fill node visuals for predefined nodes
+		node_mgr()->applyFunction([] (ContentFeatures &f) {
+			if (!f.visuals)
+				f.visuals = new NodeVisuals(&f);
+		});
 	}
 
 	MeshMakeData makeSingleNodeMMD(bool smooth_lighting = true)
@@ -57,14 +69,15 @@ public:
 		itemdef.name = "test:" + name;
 		itemdef.description = name;
 
-		ContentFeatures f;
+		CContentFeatures f;
+		f.visuals = new NodeVisuals(&f);
 		f.name = itemdef.name;
 		f.drawtype = NDT_NORMAL;
-		f.solidness = 2;
+		f.visuals->solidness = 2;
 		f.alpha = ALPHAMODE_OPAQUE;
 		for (TileDef &tiledef : f.tiledef)
 			tiledef.name = name + ".png";
-		for (TileSpec &tile : f.tiles)
+		for (TileSpec &tile : f.visuals->tiles)
 			tile.layers[0].texture_id = texture;
 
 		return registerNode(itemdef, f);
@@ -77,10 +90,11 @@ public:
 		itemdef.name = "test:" + name + "_source";
 		itemdef.description = name;
 
-		ContentFeatures f;
+		CContentFeatures f;
+		f.visuals = new NodeVisuals(&f);
 		f.name = itemdef.name;
 		f.drawtype = NDT_LIQUID;
-		f.solidness = 1;
+		f.visuals->solidness = 1;
 		f.alpha = ALPHAMODE_BLEND;
 		f.light_propagates = true;
 		f.param_type = CPT_LIGHT;
@@ -91,7 +105,7 @@ public:
 		f.liquid_alternative_flowing = "test:" + name + "_flowing";
 		for (TileDef &tiledef : f.tiledef)
 			tiledef.name = name + ".png";
-		for (TileSpec &tile : f.tiles)
+		for (TileSpec &tile : f.visuals->tiles)
 			tile.layers[0].texture_id = texture;
 
 		return registerNode(itemdef, f);
@@ -104,10 +118,11 @@ public:
 		itemdef.name = "test:" + name + "_flowing";
 		itemdef.description = name;
 
-		ContentFeatures f;
+		CContentFeatures f;
+		f.visuals = new NodeVisuals(&f);
 		f.name = itemdef.name;
 		f.drawtype = NDT_FLOWINGLIQUID;
-		f.solidness = 0;
+		f.visuals->solidness = 0;
 		f.alpha = ALPHAMODE_BLEND;
 		f.light_propagates = true;
 		f.param_type = CPT_LIGHT;
@@ -118,8 +133,8 @@ public:
 		f.liquid_alternative_flowing = "test:" + name + "_flowing";
 		f.tiledef_special[0].name = name + "_top.png";
 		f.tiledef_special[1].name = name + "_side.png";
-		f.special_tiles[0].layers[0].texture_id = texture_top;
-		f.special_tiles[1].layers[0].texture_id = texture_side;
+		f.visuals->special_tiles[0].layers[0].texture_id = texture_top;
+		f.visuals->special_tiles[1].layers[0].texture_id = texture_side;
 
 		return registerNode(itemdef, f);
 	}
